Split out thread pool tests into threadpool_test
